requirements.txt
discord.py>=2.3.0
python-binance>=1.0.19
python-dotenv>=1.0.0
